+++ app/base/__init__.py
@@ -1,3 +1,4 @@
+"""Blueprint for base endpoints."""
 from flask import Blueprint
 
 base_blueprint = Blueprint("base_blueprint", __name__, template_folder="templates")