+++ app/menus/__init__.py
@@ -1,3 +1,4 @@
+"""Blueprint for menus endpoints."""
 from flask import Blueprint
 
 menus_blueprint = Blueprint("menus_blueprint", __name__, template_folder="templates")