+++ app/tests/test_menus/test_core/test_menus_utils.py
@@ -403,7 +403,7 @@ class TestPatterns:
             "día: 25 de abril de 2020 (viernes)",
         ),
         ("día : 25 de abril de 2020 ( viernes )", "día: 25 de abril de 2020 (viernes)"),
-        ("día: invalid", None)
+        ("día: invalid", None),
     )
 
     @pytest.mark.parametrize("string, expected_sub", fix_dates_patterns_3_data)
@@ -417,14 +417,14 @@ class TestPatterns:
         else:
             assert pattern_match is None
 
-    fix_content_pattern_1 = (
+    fix_content_pattern_1_data = (
         ("cocido\ncompleto", "cocido completo"),
         ("fruta\ndía:", None),
         ("mantequilla, mermelada\ncomida:", None),
         ("mantequilla, mermelada\ncomida;", None),
     )
 
-    @pytest.mark.parametrize("string, expected_sub", fix_content_pattern_1)
+    @pytest.mark.parametrize("string, expected_sub", fix_content_pattern_1_data)
     def test_fix_content_pattern_1(self, string, expected_sub):
         real_sub = Patterns.fix_content_pattern_1.sub(r"\1 \2", string)
         pattern_match = Patterns.fix_content_pattern_1.search(string)
@@ -435,7 +435,7 @@ class TestPatterns:
         else:
             assert pattern_match is None
 
-    fix_content_pattern_2 = (
+    fix_content_pattern_2_data = (
         ("    ", " "),
         ("\t \t\t \t", " "),
         ("\t", None),
@@ -444,7 +444,7 @@ class TestPatterns:
         ("hola que tal estás", None),
     )
 
-    @pytest.mark.parametrize("string, expected_sub", fix_content_pattern_2)
+    @pytest.mark.parametrize("string, expected_sub", fix_content_pattern_2_data)
     def test_fix_content_pattern_2(self, string, expected_sub):
         real_sub = Patterns.fix_content_pattern_2.sub(r" ", string)
         pattern_match = Patterns.fix_content_pattern_2.search(string)
@@ -455,12 +455,12 @@ class TestPatterns:
         else:
             assert pattern_match is None
 
-    fix_content_pattern_3 = (
+    fix_content_pattern_3_data = (
         ("1er plato: sardinas postre: manzana", "1er plato: sardinas\npostre: manzana"),
         ("el postre será barato", None),
     )
 
-    @pytest.mark.parametrize("string, expected_sub", fix_content_pattern_3)
+    @pytest.mark.parametrize("string, expected_sub", fix_content_pattern_3_data)
     def test_fix_content_pattern_3(self, string, expected_sub):
         real_sub = Patterns.fix_content_pattern_3.sub(r"\1\n\2", string)
         pattern_match = Patterns.fix_content_pattern_3.search(string)
@@ -468,5 +468,31 @@ class TestPatterns:
         if expected_sub:
             assert pattern_match is not None
             assert real_sub == expected_sub
+        else:
+            assert pattern_match is None
+
+    pattern_lunch_data = (
+        ("comida:", True),
+        ("comida;", True),
+        ("invalid", False),
+    )
+
+    @pytest.mark.parametrize("string, expected_match", pattern_lunch_data)
+    def test_pattern_lunch(self, string, expected_match):
+        pattern_match = Patterns.pattern_lunch.search(string)
+
+        if expected_match:
+            assert pattern_match is not None
+        else:
+            assert pattern_match is None
+
+    pattern_dinner_data = (("cena:", True), ("cena;", True), ("invalid", False))
+
+    @pytest.mark.parametrize("string, expected_match", pattern_dinner_data)
+    def test_pattern_dinner(self, string, expected_match):
+        pattern_match = Patterns.pattern_dinner.search(string)
+
+        if expected_match:
+            assert pattern_match is not None
         else:
             assert pattern_match is None