+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * New argument for `get_menus_urls`: `request_all`. If True, it will return all menus urls.
 * Add downloader class with retry control: `Downloader`
 * Add footer with contact info.
+* Use bootstrap interface in `/notifications`, `/version` and `/feedback`.
 
 ### Changed
 * Changed test data storage system.